misc: mediatek: teei: 280: tz_driver: Add forgotten unsigned long cast
Fixes the following warning during build time:

  drivers/misc/mediatek/teei/280/tz_driver/backward_driver.c: In function '__reetime_handle':
  drivers/misc/mediatek/teei/280/tz_driver/backward_driver.c:203:48: warning: passing argument 2 of 'Invalidate_Dcache_By_Area' makes integer from pointer without a cast
    Invalidate_Dcache_By_Area((unsigned long)ptr, ptr + 4);
                                                  ^
  In file included from drivers/misc/mediatek/teei/280/tz_driver/backward_driver.c:23:0:
  drivers/misc/mediatek/teei/280/tz_driver/include/teei_id.h:168:20: note: expected 'long unsigned int' but argument is of type 'void *'
   static inline void Invalidate_Dcache_By_Area(unsigned long start,
                      ^

Additionally under vfs_thread_function() virt_addr is an unsigned long
to begin with, so there's no reason to cast here.

+++ drivers/misc/mediatek/teei/280/tz_driver/backward_driver.c
@@ -200,7 +200,7 @@ int __reetime_handle(struct service_handler *handler)
 	int time_type = 0;
 
 	ptr = handler->param_buf;
-	Invalidate_Dcache_By_Area((unsigned long)ptr, ptr + 4);
+	Invalidate_Dcache_By_Area((unsigned long)ptr, (unsigned long)ptr + 4);
 	time_type = *((int *)ptr);
 	if (time_type == GET_UPTIME) {
 		get_monotonic_boottime(&tp);
@@ -266,8 +266,7 @@ static unsigned long buff_vaddr;
 int vfs_thread_function(unsigned long virt_addr,
 			unsigned long para_vaddr, unsigned long buff_vaddr)
 {
-	Invalidate_Dcache_By_Area((unsigned long)virt_addr,
-					virt_addr + VFS_SIZE);
+	Invalidate_Dcache_By_Area(virt_addr, virt_addr + VFS_SIZE);
 	daulOS_VFS_share_mem = (unsigned char *)virt_addr;
 #ifdef VFS_RDWR_SEM
 	up(&VFS_rd_sem);
